Codec2: support format change case for GBM buffer pool
For 10bit decoding, the initial pixel format is set to
8bit pixel format by Codec2 HAL. After port reconfig,
Codec2 HAL gets right pixel format, which is 10bit one.
Then 10bit pixel format is set by Codec2 HAL. Therefore,
there's a pixel format change here. Need to combine it
with width and height as a identity when manage buffers.

Change-Id: I8a84c98fc703f63fa92446b28adda8a71e8c986a
Signed-off-by: Miaomiao Chen <quic_miaochen@quicinc.com>