Codec2: add a parameter "isC2D" to the acquireExtBuffer() function
To support "external buffer + interlaced stream" case, both decoder and
C2D will acquire buffers from the same downstream buffer pool. Since the
number of buffers is limited, we need to limit the number of buffers
that decoder can take to avoid there is no buffer when C2D is trying to
acquire, therefore we add this parameter to distinguish who is taking
the buffer.

Change-Id: Ibbc3c56002a048b89f1502d98da897070ab58c4e
Signed-off-by: Shiming Dong <quic_shimdong@quicinc.com>