Update parcel data pointer after realloc with size 0
If restartWrite is called with desired size of 0, mData will be
reallocated to size 0. This frees the memory and returns a null pointer.
When this happends we need to update the stored data pointer and
capacity otherwise we will crash with a double free when the object is
desctructed.

Bug: 157066561
Test: build POC included in bug. 'adb push binderMemSafety
/data/local/tmp && adb shell /data/local/tmp/binderMemSafety'. Reproduce
the crash without this change, then verify no crash with this change.
This is also being added to STS.
Ran 'atest -p' for binder tests.

Git-commit: 4a0a55e0b68e34f411e436b19e3997a81078cdeb
Git-repo: https://android.googlesource.com/platform/frameworks/native
Change-Id: Iaf56c6f230588d7d89dc56f801e00f0ac6542a80
Signed-off-by: pengy <quic_pengy@quicinc.com>